Futher code cleanup and fixes
Fixes code consistancy and cleanup so it doesn't have unintended warnings
and responses from linters.

Remove int cast from quote to prevent tidy-clang from complaining
(gcc doesn't think this narrowing cast is an issue though).

Add a missing unused argv marker.

Update use_after_free_2_safe.c to match use_after_free_2_unsafe.c

+++ overflow/quote_overflow_2_unsafe.c
@@ -100,9 +100,9 @@ void quote_s(const char *const str, size_t length, char *const buf) {
                 j++;
                 buf[j] = 'x';
                 j++;
-                buf[j] = (((int) c & 0xf0) >> 4) + '0'; // hex of most sig digit
+                buf[j] = ((c & 0xf0) >> 4) + '0'; // hex of most sig digit
                 j++;
-                buf[j] = ((int) c & 0x0f) + '0'; // hex of least sig digit
+                buf[j] = (c & 0x0f) + '0'; // hex of least sig digit
                 j++;
             }
         }
